Eliminate compiler warnings w/ Solaris Studio

+++ simd/i386/jsimd.c
@@ -32,7 +32,7 @@
 #define IS_ALIGNED_SSE(ptr)  (IS_ALIGNED(ptr, 4)) /* 16 byte alignment */
 #define IS_ALIGNED_AVX(ptr)  (IS_ALIGNED(ptr, 5)) /* 32 byte alignment */
 
-static unsigned int simd_support = ~0;
+static unsigned int simd_support = (unsigned int)(~0);
 static unsigned int simd_huffman = 1;
 
 /*
